requirements.txt
# Core LlamaIndex
llama-index-core>=0.10.0

# Ollama integrations
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0

# ChromaDB vector store
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0

# PDF reader
llama-index-readers-file>=0.1.0
pymupdf>=1.23.0

# UI
streamlit>=1.35.0
